fix(abi): temporaryInformation and significantDigits are uint8_t, not u16/u32
Cross-owner-consistency scan follow-up. Both C globals are uint8_t (c47.h:423,
396) and most owners bind them as u8, but a few had the wrong width -- a size
mismatch that reads/writes past the 1-byte symbol into adjacent globals:

- temporaryInformation: u16 -> u8 in frontier_elec, calc_state_io_flow, and
  calc_state_save. These WRITE `temporaryInformation = TI_*`; as a u16 that stores
  two bytes over a one-byte symbol, zeroing the neighbouring global's byte on
  firmware (little-endian). The parity passed only because it re-reads the low
  byte, not the clobbered neighbour.
- significantDigits: u32 -> u8 in frontier_screen. It READS `significantDigits <=
  16`; as a u32 the three high bytes come from adjacent globals, so the guard can
  read a huge value and wrongly fail even when the real 1-byte value is <= 16.

The calc_state fake already declares uint8_t, so the fix also aligns the owners
with the test double.

Verified: zig build sim and zig build test exit 0 (9674 testSuite),
calc_state_parity green, fmt and idiom-ratchet clean.

+++ zig_src/frontier/frontier_elec.zig
@@ -85,7 +85,7 @@ const TripleRegZ3_98: calcRegister_t = 98;
 // Globals and function externs
 // ---------------------------------------------------------------------------
 extern var ctxtReal39: realContext_t;
-extern var temporaryInformation: u16;
+extern var temporaryInformation: u8;
 
 extern fn getRegisterDataType(regist: calcRegister_t) u32;
 extern fn getRegisterDataPointer(regist: calcRegister_t) ?*anyopaque;
